DBZ-566 Make embedded engine implementation internal
